All of the steps from 0 to online.
This is the set of notes from the start to having http://calendar.acaird.com
working.

Also, this commit is the first one after the repo was made private, so all of
the details in here aren't publicly available.  But it does mean I need to
write a blog post.