Add ICS and PDF file diff'ing
The actual test - do the results match the answers.

diff'ing PDF files is a weird thing to do, but the ReportLab PDF Library seem
consistent and only makes differences in the CreationTime and the checksum
lines.  Sadly, I had to call out to `diff` to get this; I couldn't find a
Python library that would help me with this.

It's also worth noting that the current commit of the `answers/*` files is
actually wrong. :(